requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
pydantic-settings>=2.0.0
python-multipart>=0.0.5
aiohttp>=3.8.0
PyYAML>=5.4.0
opencv-python>=4.8.1.78
numpy>=1.26.2
ultralytics>=8.0.0,<8.4.0
redis[hiredis]>=5.0.1
aioredis>=2.0.0
psutil==5.9.6
gputil==1.4.0
loguru>=0.7.0
scipy>=1.11.0
GPUtil>=1.4.0
httpx>=0.24.0
paho-mqtt>=1.6.1